pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "timelapse"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "picamera2>=0.3.19",
    "astral>=3.2",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
mqtt = ["paho-mqtt>=2.0"]
dev = ["pytest>=8.0"]

[project.scripts]
timelapse = "timelapse.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that cross component boundaries or use real external tools (ffmpeg)",
]
